improved handling of C types
The Val type now carries type and pointer information.

Product and sum types (their structs) are always const restrict
pointers, since they're read-only and binders are never reassigned
except in the forthcoming sum elimination implementation (we'll just
special case that).

The eval function will never return a pointer; if the thing returned is
a pointer it will dereference it as many times as necessary.