no GHC Typeable use; use const almost everywhere
for sums and products, all members of structs and unions are const.
Any members which are also compound types are const restrict pointers.
This should be OK since we indeed never update these values. The only
update we ever do is for the result value in a sum elimination, which
starts uninitialized.