experimental feature: type normalization
The idea is that the C backend will always normalize types before
working with them. This ensures that nested products and sums are
flattened, void appears nowhere within a compound type (it annihilates
a conjunction and is identity in a disjunction), units do not appear in
products, etc. etc.

From here, we can get nice optimizations such as the simple enum
optimization: a sum where each summand is unit can be represented as an
enum. We can even make it so that representing booleans as 1 + 1 is
consistent with the built-in C logical operators.