sum/product reps do not use pointers
at first glance this looks like a horrible idea: now we have to copy the
entire compound structure every time we make a composite containing a
composite. But on the other hand, we now have definitions which are
suitable for easily declaring values of composite types with static
storage, which will be essential when dealing with streams in the
future.

Since the types in the object language are all finite and non-recursive,
the C type declarations are all guaranteed to come out in a legit order,
with no forward declarations necessary.

It should be possible to include, as an optimization, a treatment of
compound types by way of reference (as they were before). The point
however is that this optimized style should not be considered the basic
idea. That's to say, we shouldn't start there and try to write out full
types from these only when they are needed.