compound types by reference option
Not super happy with the implementation but it's a start.

The idea is that the generated C should always give the same result,
whether we use sharing or not, but maybe the use of sharing will make it
more performant (not clear whether state-of-the-art C compilers can make
these optimizations, even if everything is const).