non-shared and shared representation basic working
This is probably still buggy but it works in basic cases: a compound
type in "shared" representation can be converted into one which doesn't
have any internal pointers, and therefore is suitable for, say,
returning from the eval function and holding in some static memory area.
This will be key when streams are defined.