extern outputs in the C interpreter
For symmetry with the extern input style. Give a stream, and its value
will be assigned to the extern of that name at each evaluation of the
system.

This is probably not the IO model that we want to work with (would maybe
prefer function calls?) but it's at least good enough to start doing
some "real" applications.