Expr is not a monad, Exec monad defined
The latest in my attempt to sort out exactly what the form should be.
There are still 3 main parts:
- edsl
- value
- context (monad)
An expression is an AST with naming/binding, and with two parts: the
edsl and the value, representing respectively the "interpreter-agnostic"
part and the "interpreter-specific" part (already-computed values like
C externs for instance).

Then there is the Exec monad which introduces a context for computing
interpreter-specific parts, and into which expressions over that
interpreter's value type may be brought.

The motivating example for this is the C interpreter, in which the
construction of inputs and ouputs (extern declarations) must be ordered
(in the CodeGen monad). It wouldn't make sense to put "externOutput"
into the AST itself, since that loses the uniqueness and order; but it
does make sense to give back a C-specific value (C.Stream) within the
Exec monad over CodeGen, and to allow for that value to be used within
an expression.