object language has Program functor
It's similar to the role that IO plays in Haskell. The main motivation
for this is to ensure that the recursive knot definition function cannot
itself create more recursive knots. Since its type is always a product
of varyings to a product of varyings, this holds, for there is no way to
eliminate the Program type.

One weird thing about Program is that it takes any meta-language type
over the object language, whereas our other functors `Varying n` take
only point types, and therefore are not endofunctors.

Something to look into in the future: endofunctors / monads within
point, such as Maybe. Is it feasible to give a monadic-style interface
(do notation, most desirably) without requiring that the "functor" have
arrows in its domain? i.e. in a setting where `Maybe (s :-> t)` is
ill-typed?