let bindings only for constants
Having a way to express sharing in the EDSL has always been a
requirement. And allowing for this sharing to be expressed _over_
meta-language products is key. That's the whole reason why the
representation is factored into f and val. But having such a general
notion of let binding makes it difficult to come up with a C
code generation interpreter. A let binding should correspond to a C
assignment, so that the value can be shared by storing it on the stack.
But if the value is a varying, then what? Store the value for every
prefix on the stack? Or just the current frame? The former is probably a
waste of space and the latter is confusing.

Let-binding of constants is comparatively easy, and it ought to be
enough to express sharing even over varyings, by judiciously using
let-bindings in the functions which are lifted/mapped over varyings.